Refine JavaDoc and move Log filters to 'output'
