+++ src/main/java/fr/xephi/authme/output/ConsoleFilter.java
@@ -1,4 +1,4 @@
-package fr.xephi.authme;
+package fr.xephi.authme.output;
 
 import java.util.logging.Filter;
 import java.util.logging.LogRecord;
@@ -7,20 +7,9 @@ import java.util.logging.LogRecord;
  * Console filter Class
  *
  * @author Xephi59
- * @version $Revision: 1.0 $
  */
 public class ConsoleFilter implements Filter {
 
-    public ConsoleFilter() {
-    }
-
-    /**
-     * Method isLoggable.
-     *
-     * @param record LogRecord
-     *
-     * @return boolean * @see java.util.logging.Filter#isLoggable(LogRecord)
-     */
     @Override
     public boolean isLoggable(LogRecord record) {
         try {
