+++ src/test/java/fr/xephi/authme/output/Log4JFilterTest.java
@@ -1,4 +1,4 @@
-package fr.xephi.authme;
+package fr.xephi.authme.output;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
@@ -12,8 +12,6 @@ import org.mockito.Mockito;
 
 /**
  * Test for {@link Log4JFilter}.
- * @author Gabriele
- * @version $Revision: 1.0 $
  */
 public class Log4JFilterTest {
